requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
aiohttp==3.8.6
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
urllib3==2.0.9
